SciChart version 6.5.0.11860

+++ Examples/SciChart.Examples/Examples/PerformanceDemos2D/DigitalAnalyzer/Common/ChannelGenerationHelper.cs
@@ -24,16 +24,18 @@ namespace SciChart.Examples.Examples.PerformanceDemos2D.DigitalAnalyzer.Common
                 digitalData[i + 1 >= count ? i : i + 1] = bit;
             }
 
+            // Provide additional info about expected data to avoid runtime checks
+            // There are frequent peaks at 0,1
             var args = new UniformDataDistributionArgs<byte>(false, 0, 1);
             var dataSeries = new UniformXyDataSeries<byte>(xStart, xStep, digitalData, args);
 
             return new ChannelViewModel(dataSeries, new DoubleRange(-0.5, 1.5), index, $"Channel {index}") { IsDigital = true };
         }
 
-        public ChannelViewModel GenerateAnalogChannel(double xStart, double xStep, double[] analogData, int index = 0)
+        public ChannelViewModel GenerateAnalogChannel(double xStart, double xStep, float[] analogData, int index = 0)
         {
             UniformDataManager.GenerateAnalogData(analogData);
-            var dataSeries = new UniformXyDataSeries<double>(xStart, xStep, analogData);
+            var dataSeries = new UniformXyDataSeries<float>(xStart, xStep, analogData);
 
             return new ChannelViewModel(dataSeries, new DoubleRange(-1.5, 1.5), index, $"Channel {index}") { IsDigital = false };
         }
