+++ Examples/SciChart.Examples/Examples/PerformanceDemos2D/DigitalAnalyzer/Common/UniformDataManager.cs
@@ -4,7 +4,7 @@ namespace SciChart.Examples.Examples.PerformanceDemos2D.DigitalAnalyzer.Common
 {
     public static class UniformDataManager
     {
-        public static void GenerateAnalogData(double[] buffer)
+        public static void GenerateAnalogData(float[] buffer)
         {
             var count = buffer.Length;
             var freq = count / 100;
@@ -15,7 +15,7 @@ namespace SciChart.Examples.Examples.PerformanceDemos2D.DigitalAnalyzer.Common
             for (int i = 0, j = 0; i < count; i++, j++)
             {
                 var wn = 2 * Math.PI / (count / (double)freq);
-                buffer[i] = amp * Math.Sin(j * wn + phase);
+                buffer[i] = (float)(amp * Math.Sin(j * wn + phase));
             }
         }
     }
